feat: :sparkles: Watcher
