+++ 03_basic_v3/src/main.js
@@ -8,6 +8,7 @@ import './style.css';
 // import App from './App06.vue';
 // import App from './App07.vue';
 // import App from './App08.vue';
+// import App from './App09.vue';
 import App from './App.vue';
 
 createApp(App).mount('#app');
